parse timestamp as int

+++ ios/Classes/LinearAccelerationStreamHandler.swift
@@ -44,7 +44,7 @@ public class LinearAccelerationStreamHandler : NSObject, FlutterStreamHandler {
                 values.y,
                 values.z
             ]
-            SwiftFlutterSensorsPlugin.notify(sensorId: LinearAccelerationStreamHandler.SENSOR_ID, sensorData: dataArray, timestamp: deviceMotion.timestamp * 1000, eventSink: eventSink)
+            SwiftFlutterSensorsPlugin.notify(sensorId: LinearAccelerationStreamHandler.SENSOR_ID, sensorData: dataArray, timestamp: Int(deviceMotion.timestamp * 1000), eventSink: eventSink)
         })
     }
